Added support for map/layer mask
A custom "Mask" property on a layer or a "foregroundXmask" property on a
map can now be used in combination with the SMSG_MAP_MASK to dynamically
disable certain map layers from the server.

Feature previously seen on ManaPlus and implemented for Mana client for
compatibility.

Also added a ResourceRef class for automating the Resource reference
counting.

Closes #44