Allow mouse focus click-through
This is generally expected on Linux, where it feels annoying that the
initial click just focuses the window.

On macOS it's a little bit different, since applications can be specific
about which elements are directly clickable and which aren't, but in
general it feels better to me when the UI can be interacted with
directly.