Use SDL2 support for color and system mouse cursors
This way the cursor is not limited by the framerate nor affected by
input lag. Also, when custom cursor is disabled, a few different system
cursors are now used instead.

It also avoids an issue on Wayland, where hiding the cursor (as done to
render our own one) would cause the cursor to get locked within the
window.

On macOS it fixes two cursors being visible when hovering the window
while it is in the background.

The cursor can unfortunately no longer gently fade away.