Resolved issues with winver.h
The configured file wasn't actually being used, because it was not in
the include path. Instead, a checked in version was used.

In addition, winver.h is an existing Windows header, which was
conflicting with the generated file. Instead, now mana.rc itself is
configured. We didn't actually need winver.h anywhere else, anyway.