Fixed choosing preferred language on Windows
The linker could not find _nl_locale_name_default. Maybe it could be
fixed by linking to another library, but it would likely not yield the
preferred behavior. Instead, we now use GetUserPreferredUILanguages to
set the LANG environment variable with one or more languages and let
gettext choose the best fitting available translation.

Also fixed the directory from which the translations are loaded to match
the installation directory,