CMake: Use FindIntl
FindIntl was added with CMake 3.2 and can be used to find the Gettext
libintl headers and libraries.

I've removed special handling of Solaris, which explicitly added linking
to 'intl' and added '/usr/local/lib' include path. It can be easily
restored if necessary, once somebody tries to compile Mana on Solaris.