Fixed the scaling of the custom cursor
It appears that either SDL or the system is already taking care of
scaling the cursor to the display scale (observed on both macOS and
Wayland). Hence, we should only scale the cursor by the user scale and
not by both.

Follow-up to 6eca1b485dba7355d827745284ed2f0072f9e370.