Fixed being popup getting stuck under the mouse
Normally the popup follows the mouse, but when the mouse is moved down
such that it is on top of the popup, the popup would seem to get stuck.

When the BeingPopup is a mouse listener, the default behavior of Popup
hiding the BeingPopup when the mouse moves over them kicks in.

Hiding of the BeingPopup by Popup and Window when they are hovered in
general is a workaround which might be necessary because there is
currently no way for the Viewport to determine whether the mouse is
above the map view or over a piece of UI.