Added online player list to Social window
The online list refreshes every 18 seconds, which matches ManaVerse 
behavior. It's not ideal, but to improve this would mean diving into
TMWA.

The client version was bumped to 8 to get a SMSG_ONLINE_LIST reply. 
Further changes needed related to the client version are tracked by #71. 

This also changes the TabbedArea to take into account the frame size for
its tab widgets, to make sure those frames are not clipped by the
TabbedArea widget (as happened in the Social window).

The horizontal scroll bar is now disabled in all social tabs, with the
vertical one appearing only when necessary.

Closes #61