Added support for hit/miss sounds on equipment for all players
Previously only the local player's weapon "strike" sound would play,
regardless of hit or miss. Now the sound is played in response to the
SMSG_BEING_ACTION message, so it can be played for all players.

Also added alias "miss", which is used by TMW in some places.

Finally, when no weapon is equipped, it falls back to the sounds defined
on the racesprite item (fixes punch sound on TMW).

Closes #68