Fixed ConfigOptionChanged events
With the statically typed config we no longer get an event for each
changed config value. Where relevant, this is now done through
`setConfigValue`.

The `Event` now uses a `std::any`, which for `ConfigOptionChanged`
events is set to the changed `Config` member. This allows for a
type-safe check on which config value was changed.