Replaced ParticleContainer with a ParticleHandle
The main feature of the ParticleContainer seemed to be to provide
persistent access to particles, even though they are generally owned by
their parent particle. This allowed particles to be moved along with the
ActorSprite and to be killed explicitly when they were no longer wanted.

The ParticleHandle now takes care of disabling automatic deletion and of
killing the particle along with the handle. It allows the particle
references to be stored in whatever container is suitable, like a vector
for general ActorSprite particle effects, separate vectors for equipment
based particles and finally a map for status effect particles.

Moving the particles along with the ActorSprite does need to happen more
explicitly now, but overall it's a nice cleanup.