Fixed status effect icon positions
The status effect icons are implemented as sprites. Traditionally, since
sprites were rendered on the map, they were by default centered
horizontally on a tile and aligned to the bottom of the tile. The
necessary offsets were part of the animation frames.

This was changed in ae4aea451d6e2c44b273c963026c4fd697568f79 and
0248b2f58c783f21b7e68a20deadc0a3f942c869, which moved this offset to be
applied at a later stage. At that point, client data was changed to
remove this offset from the status effect icons, where it wasn't
meaningful.

Unfortunately, those client data changes were later reverted because the
M+ fork of the client still applies the above offsets to the animation
frames. Hence, to fix the status icon sprite positions, Mana client now
applies the same offsets when rendering the status icons.

Closes https://git.themanaworld.org/mana/mana/-/issues/95