Cleanup Download code, fixing SDL_Thread leaking
The download thread was setting itself to nullptr (d->mThread = nullptr)
in a number of locations. This caused a later call to SDL_WaitThread to
be unable to perform cleanup.

This reverts most of 1eb02f83a5d3895e4e18db30ea10d88da94ba4c0 (including
making Download::cancel no longer blocking), but keeps the necessary
waiting for the thread to finish before freeing the memory buffer in
~UpdaterWindow(), which might have been the bug fixed by that change.

Fixed removal of downloaded .part file when its checksum failed. It
trying to remove the file without .part appended instead.

Fixed download progress indication in ServerDialog to not be reversed,
though this is rarely visible due to the server list being so small.

Fixed reporting of curl error.