Introduced spriteOffsetY setting to replace half tile hack
For historical reasons, sprites get a 16 pixel offset by default, which
is used to position them correctly on their tile.

When pixel-based movement was added, actors are now positioned on the
middle of a tile and the built-in offset was removed. Servers that use
sprites that do not rely on this offset can now set the spriteOffsetY
option to 0.