Use a consistent naming style for enum class values
Sometimes I've used CamelCase and sometimes SNAKE_CASE for these values.
Since "enum class" values are always prefixed with the enum name, which
uses CamelCase, I find it more fitting to use it for the values as well.

Also fixes compilation on Windows where 'ERROR' was conflicting with a
define.