Added the Jewelry theme
This theme was created by Francesco Miglietta in 2012-2013 for ManaPlus
and is hereby ported to the new XML theme definition supported by the
Mana client.

A few things are still missing:

* Use of special window skins for some windows (minimap, chat, shortcuts)
* Control of text format in chat window and popups
* Theming the equipment box icons
* Support for fixed-size scroll bar handles
* Not all colors set by the theme are supported yet

There is also still no way for the user to choose the theme in the Setup
dialog. For now, the theme can be enabled by manually editing the config
file to contain:

    <option name="theme" value="jewelry"/>