Avoid pixel format conversions when applying dye
The images returned by IMG_Load_RW appear to usually be in the format
SDL_PIXELFORMAT_ABGR8888. As such, adjust the code performing the dye to
operate on this format rather than always converting the surfaces to
SDL_PIXELFORMAT_RGBA8888.

This appears to make the loading of dyed images about 30% faster.