GitHub Actions: Added macOS build
arm64 and x86_64 are built separately, then merged into a universal app
using lipo. The creation of the DMG is now handled by the create-dmg
tool rather than CPack.

On macOS Arm, Homebrew uses a different prefix (/opt/homebrew), which
needed to be added to the fixup_bundle call.

ccache was enabled to speed up the builds.