Delay showing of the window to avoid startup flicker
Otherwise, at least on X11, the window is (visibly) re-created once the
SDL renderer is created. Also the window icon can be seen changing on
startup. Showing the window only after all related setup has been
completed avoids this.

On X11 this appears to have the added benefit of putting a "windowed
fullscreen" window on the active screen (in terms of where the mouse is)
rather than always using the primary screen.

On Windows, showing the window only after the initial paint avoids the
window initially appearing with its contents entirely black.

The window is still immediately shown in the following cases:

* In exclusive fullscreen mode, because on Windows this caused the
  initial show of a fullscreen window to take much longer and on X11 it
  could cause the resolution of the primary screen to be changed while
  maximizing the Mana window on the active screen.

* In "windowed fullscreen" mode when the legacy OpenGL renderer is used,
  where it was causing an additional flicker on Windows.