Added support for the hidden/neutral gender
TMWA supports a "neutral" gender, which in Mana client maps to "hidden". In
The Mana World, it means a player character is not displayed as clearly male
or female, using a dedicated base sprite with more underwear.

It is not possible to choose the neutral gender on account registration or
character creation. In TMW, these choices are not actually relevant anyway
and could probably be removed.

Fixes #114