Work around player invisibility caused by "unspecified" gender
The "Unspecified" gender was causing some older players to be invisible
since no sprites were displayed at all. I think this is a bug in TMWA,
which should never send SEX::UNSPECIFIED for player characters, but
while this bug exists we can make sure to map this to Gender::Neutral
(renamed from "Hidden").

The Gender::Unspecified value was removed entirely because it never
makes sense for beings. One "neutral" value should be enough.