Reduced the size of the character select window
The labels for showing character level and money were moved into a hover
popup. The various account management buttons were moved into a panel on
the right.

These changes are especially meant to reduce the height of the window so
it can comfortably show two rows of characters at low resolutions.