Use online player list for name autocomplete
Now the we have the list of all online players, use this to expand the
scope of autocompletion. Keep using local players as fallback for when
the online list is not available.

Added ChatHandler::getOnlinePlayerNames and implemented this for
TmwAthena::ChatHandler. Cleaned up autocomplete logic and derived
PlayerNPCNamesLister from PlayerNamesLister.