Use the theme to draw the health bars
The health bars now have a distinct look in each of the three supported
UI themes. The health bars also adjust to the monster size now.

The positioning of the health bar was adjusted to be always aligned to
the bottom of the name, with the name moving up a little.

The targeted monster now shows its health bar also when it isn't
damaged.